# .git-rewrite/t/apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Lekhaly API
  version: 0.1.0
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
  /auth/login:
    post:
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [companyCode, email, password]
              properties:
                companyCode: { type: string }
                email: { type: string }
                password: { type: string }
                totpCode: { type: string }
                deviceId: { type: string }
                deviceLabel: { type: string }
                rememberDevice: { type: boolean }
      responses:
        "200": { description: OK }
  /auth/register:
    post:
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [companyCode, companyName, name, email, password]
              properties:
                companyCode: { type: string }
                companyName: { type: string }
                name: { type: string }
                email: { type: string }
                password: { type: string }
      responses:
        "200": { description: OK }
...
  /v1/auth/login:
    post:
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [companyCode, email, password]
              properties:
                companyCode: { type: string }
                email: { type: string }
                password: { type: string }
                totpCode: { type: string }
                deviceId: { type: string }
                deviceLabel: { type: string }
                rememberDevice: { type: boolean }
      responses: { "200": { description: OK } }
  /v1/auth/register:
    post:
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [companyCode, companyName, name, email, password]
              properties:
                companyCode: { type: string }
                companyName: { type: string }
                name: { type: string }
                email: { type: string }
                password: { type: string }
      responses: { "200": { description: OK } }
...